Fix hanging request bug
In cases where we had an error in sending (for instance, if the client
cancelled the request or timed out), then the cancel() method for the
context never actually got called until the eventual server-side timeout
(at 15 min). For some reason (which I'm not sure of), this impacted the
server's ability to make progress on any other requests.
This now explicitly calls cancel() when we get in that
situation, which seems to have addressed the issue.